simplify: remove unnecessary float64 upcasts from analysis pipeline
MD trajectories are stored in float32 by mdtraj. Empirical testing
confirms float32 precision is sufficient for all analysis operations:
RMSF (~1e-5 nm error), DCCM (~4e-6 correlation error), FES (~2e-6
kJ/mol error), all well below any physically meaningful threshold.

Removed defensive float64 upcasts from: RMSF/DCCM/delta-RMSF
(metrics.py), FES input/histogram (fes.py), TICA pre-cast
(decomposition.py, deeptime handles upcast internally), boolean mean
dtype (contacts.py, hbond.py -- numpy already defaults to float64).

Kept float64 only where external constraints require it: Numba JIT
kernel output type and numpy/deeptime internal behavior.

+++ src/mdpp/analysis/contacts.py
@@ -108,7 +108,7 @@ def compute_contact_frequency(
         scheme=scheme,
         periodic=periodic,
     )
-    frequency = np.mean(distances < cutoff_nm, axis=0, dtype=np.float64)
+    frequency = np.mean(distances < cutoff_nm, axis=0)
     return np.asarray(frequency, dtype=resolved), np.asarray(pairs, dtype=np.int_)
 
 
@@ -165,7 +165,7 @@ def compute_native_contacts(
     native_pairs = pairs[native_mask]
     native_distances = distances[:, native_mask]
     resolved = resolve_dtype(dtype)
-    fraction = np.mean(native_distances < cutoff_nm, axis=1, dtype=np.float64)
+    fraction = np.mean(native_distances < cutoff_nm, axis=1)
 
     return NativeContactResult(
         time_ps=trajectory_time_ps(traj, timestep_ps=timestep_ps, dtype=resolved),
